CrimeIntel/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
python-dotenv==1.0.1
apscheduler==3.10.4
scikit-learn==1.5.0
numpy==1.26.4
pandas==2.2.2
